moved BraKet enum class from engine to dedicated header
